requirements.txt
alembic==1.13.3
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
autoflake==2.3.1
bcrypt==4.2.0
billiard==4.2.1
black==24.10.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.3
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.3
fastapi-cache2==0.2.2
fastapi-cli==0.0.5
fastapi-versioning==0.10.0
flake8==7.1.1
flower==2.0.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
humanize==4.11.0
idna==3.10
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==2.1.3
orjson==3.10.10
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pathspec==0.12.1
pendulum==3.0.0
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
pyasn1==0.6.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.6.0
pydantic_core==2.23.4
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
pyright==1.1.388
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
redis==4.6.0
rich==13.9.3
rsa==4.9
sentry-sdk==2.18.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqladmin==0.20.1
SQLAlchemy==2.0.36
starlette==0.41.0
time-machine==2.16.0
tornado==6.4.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.32.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
WTForms==3.1.2
